Removed call to Style::instance().settings().initializeColors() in "instanceInit()"
It seems this call is not necessary anymore (though it was, for the v1.0.1 release, otherwise
breaking arrow background color in combobox entries).

Will revert if this causes any problem.
Will backport otherwise (and remove the corresponding initializeColors() method from
Oxygen::QtSettings), if things are working fine, after a couple of days.