Use g_mkdir() instead of different versions of mkdir().
The current code was erroneously relying on the _POSIX_C_SOURCE macro to
decide whether to use mkdir() with the signature present in POSIX-compliant
systems or with the Windows signature.

It turns out POSIX.1 expects the applications to define _POSIX_C_SOURCE to
certain values. Only glibc resorts to defining it to a value by default,
opposing what the standard says. It would then break systems that use
another libc, such as the BSDs, OS X and maybe even other Linux libc
implementations.

Work around the issue by just relying on glib's g_mkdir(), which has the
proper means to decide which mkdir() function to call. We still need to
check for whether we are on Windows, though, since some of the mode
constants are not defined on it.

REVIEW:		112402